refactor(e2e): extract shared ktor client tests into testFixtures module
- Create ktor-client-shared-tests module with java-test-fixtures plugin
- Move KtorClientTestBase to testFixtures source set
- Update ktor-client-jackson and ktor-client-kotlinx to depend on testFixtures
- Add common JVM/test configuration to root build.gradle.kts
- Eliminates "Duplicate roots detected" IDE warning
