+++ settings.gradle.kts
@@ -4,6 +4,7 @@ include(
     "end2end-tests:okhttp",
     "end2end-tests:openfeign",
     "end2end-tests:ktor",
+    "end2end-tests:ktor-client-shared-tests",
     "end2end-tests:ktor-client-kotlinx",
     "end2end-tests:ktor-client-jackson",
     "end2end-tests:spring-http-interface",
